Revert "add JsonEntityFieldMaker"
