+++ src/org/nutz/json/Json.java
@@ -14,7 +14,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.nutz.json.entity.JsonEntity;
-import org.nutz.json.impl.JsonEntityFieldMakerImpl;
 import org.nutz.json.impl.JsonRenderImpl;
 import org.nutz.lang.Files;
 import org.nutz.lang.Lang;
@@ -418,14 +417,5 @@ public class Json {
         if (defaultJf == null)
             defaultJf = JsonFormat.nice();
         Json.deft = defaultJf;
-    }
-
-    private static JsonEntityFieldMaker deftMaker = new JsonEntityFieldMakerImpl();
-    public static void setDefaultFieldMaker(JsonEntityFieldMaker fieldMaker) {
-        if (fieldMaker != null)
-            Json.deftMaker = fieldMaker;
-    }
-    public static JsonEntityFieldMaker getDefaultFieldMaker() {
-        return deftMaker;
     }
 }
