+++ test/org/nutz/json/JsonEntityTest.java
@@ -2,19 +2,14 @@ package org.nutz.json;
 
 import static org.junit.Assert.*;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.Map;
 
 import org.junit.Test;
 import org.nutz.json.entity.JsonEntity;
-import org.nutz.json.entity.JsonEntityField;
 import org.nutz.json.meta.JENObj;
 import org.nutz.lang.Lang;
 import org.nutz.lang.Mirror;
-import org.nutz.lang.Strings;
-import org.nutz.lang.inject.InjectBySetter;
 
 public class JsonEntityTest {
 
@@ -58,45 +53,6 @@ public class JsonEntityTest {
         String keyStr = Lang.concat(",", keys).toString();
         assertEquals("age,id,name", keyStr);
 
-    }
-
-    @Test
-    public void test_entity_field_maker() {
-        JENObj obj = new JENObj();
-        obj.setAge(100);
-        obj.setName("name");
-        obj.setObjId(9l);
-        Json.setDefaultFieldMaker(new AbstractJsonEntityFieldMaker() {
-            @Override
-            public JsonEntityField make(Mirror<?> mirror, Method method) {
-                return null;
-            }
-            @Override
-            public JsonEntityField make(Mirror<?> mirror, Field field) {
-                return null;
-            }
-        });
-        assertEquals("{}", Json.toJson(obj, JsonFormat.compact()));
-        Json.clearEntityCache();
-        Json.setDefaultFieldMaker(new AbstractJsonEntityFieldMaker() {
-            @Override
-            public JsonEntityField make(Mirror<?> mirror, Method method) {
-                if (method.getName().equals("setName")) {
-                    String fn = Strings.lowerFirst(method.getName().substring(3));
-                    return JsonEntityField.eval("another_name", method.getParameterTypes()[0], mirror.getEjecting(fn), new InjectBySetter(method));
-                }
-                return null;
-            }
-            @Override
-            public JsonEntityField make(Mirror<?> mirror, Field field) {
-                return JsonEntityField.eval("test_" + field.getName(), field.getType(), mirror.getEjecting(field), mirror.getInjecting(field.getName()));
-            }
-        });
-        String json = Json.toJson(obj, JsonFormat.compact());
-        assertTrue(json.contains("\"test_objId\":"));
-        assertTrue(json.contains("\"test_name\":"));
-        assertTrue(json.contains("\"test_age\":"));
-        assertTrue(json.contains("\"another_name\":"));
     }
 
 }
